# backend/internal/handler/rest/user/user.go
package user

import (
	"context"
	"encoding/json"
	stderrs "errors"

	nethttp "net/http"
	"strconv"

	apperrors "task-management-backend/internal/constant/errors"
	"task-management-backend/internal/handler/middleware"
	"task-management-backend/internal/model/db"
	"task-management-backend/internal/model/dto"
	"task-management-backend/internal/model/response"
	"task-management-backend/internal/module"
	user "task-management-backend/internal/module/user"

	"github.com/go-chi/chi/v5"
)

type Handler struct {
	module *module.Module
}

type validatable interface {
	Validate() error
}

func New(module *module.Module) *Handler {
	return &Handler{module: module}
}

func (h *Handler) Module() *module.Module {
	return h.module
}
func (h *Handler) HealthCheck(w nethttp.ResponseWriter, r *nethttp.Request) {
	response.SendSuccessResponse(w, r, nethttp.StatusOK, "🚀 workConnect service is running", nil, false)
}
func (h *Handler) Register(w nethttp.ResponseWriter, r *nethttp.Request) {
	var req dto.RegisterRequest
	if err := decodeAndValidate(r, &req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	token, user, err := h.Module().WorkConnect.Register(r.Context(), req)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusCreated, "user registered", h.authPayload(r.Context(), token, user))
}

func (h *Handler) Login(w nethttp.ResponseWriter, r *nethttp.Request) {
	var req dto.LoginRequest
	if err := decodeAndValidate(r, &req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	auth, err := h.Module().WorkConnect.Login(r.Context(), req)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "login successful", auth)
}

func (h *Handler) ChangePassword(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	var req dto.ChangePasswordRequest
	if err = decodeAndValidate(r, &req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	if err = h.Module().WorkConnect.ChangePassword(r.Context(), principal.UserID, req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "password updated", nil)
}

func (h *Handler) Me(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	user, err := h.Module().WorkConnect.GetProfile(r.Context(), principal.UserID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "profile fetched", h.profilePayload(r.Context(), user))
}

func (h *Handler) DeleteAccount(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	if err = h.Module().WorkConnect.DeleteAccount(r.Context(), principal.UserID); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "account deleted", nil)
}

func (h *Handler) UpdateProfileImage(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	var req dto.UpdateProfileRequest
	if err = decodeAndValidate(r, &req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	user, err := h.Module().WorkConnect.UpdateProfileImage(r.Context(), principal.UserID, req)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "profile image updated", h.profilePayload(r.Context(), user))
}

func (h *Handler) ListWorkers(w nethttp.ResponseWriter, r *nethttp.Request) {
	query := dto.WorkerSearchQuery{
		Category: r.URL.Query().Get("category"),
		City:     r.URL.Query().Get("city"),
		Q:        r.URL.Query().Get("q"),
		Sort:     r.URL.Query().Get("sort"),
	}

	workers, err := h.Module().WorkConnect.ListWorkers(r.Context(), query)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "workers fetched", map[string]any{"workers": workers})
}

func (h *Handler) GetWorkerProfile(w nethttp.ResponseWriter, r *nethttp.Request) {
	workerID, err := parseIDParam(r, "workerID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid worker id"))
		return
	}

	worker, err := h.Module().WorkConnect.GetWorkerDetails(r.Context(), workerID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "worker fetched", map[string]any{
		"worker": worker.Worker,
		"bio":    worker.Bio,
		"phone":  worker.Phone,
		"email":  worker.Email,
		"skills": worker.Skills,
	})
}

func (h *Handler) CreateCustomerRequest(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	var req dto.CreateServiceRequest
	if err := decodeAndValidate(r, &req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	item, err := h.Module().WorkConnect.CreateServiceRequest(r.Context(), principal.UserID, req)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusCreated, "request created", map[string]any{"request": item})
}

func (h *Handler) ListCustomerRequests(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	items, err := h.Module().WorkConnect.ListCustomerRequests(r.Context(), principal.UserID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "requests fetched", map[string]any{"requests": items})
}

func (h *Handler) GetCustomerRequest(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	requestID, err := parseIDParam(r, "requestID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid request id"))
		return
	}

	request, err := h.Module().WorkConnect.GetServiceRequestByID(r.Context(), requestID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	if request.CustomerID != principal.UserID {
		response.SendErrorResponse(w, r, apperrors.ErrForbidden)
		return
	}

	worker, err := h.Module().WorkConnect.GetWorkerDetails(r.Context(), request.WorkerID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "request fetched", map[string]any{
		"request": request,
		"worker":  worker.Worker,
		"bio":     worker.Bio,
		"phone":   worker.Phone,
		"email":   worker.Email,
		"skills":  worker.Skills,
	})
}

func (h *Handler) SubmitCustomerReview(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	requestID, err := parseIDParam(r, "requestID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid request id"))
		return
	}

	var req dto.SubmitReviewRequest
	if err = decodeAndValidate(r, &req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	if err = h.Module().WorkConnect.SubmitReview(r.Context(), principal.UserID, requestID, req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusCreated, "review submitted", nil)
}

func (h *Handler) InitiateCustomerPayment(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	requestID, err := parseIDParam(r, "requestID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid request id"))
		return
	}

	var req dto.InitiatePaymentRequest
	if err = decodeAndValidate(r, &req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	payment, err := h.Module().WorkConnect.InitiatePayment(r.Context(), principal.UserID, requestID, req)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusCreated, "payment initiated", map[string]any{"payment": payment})
}

func (h *Handler) CustomerDashboard(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	summary, err := h.Module().WorkConnect.CustomerDashboard(r.Context(), principal.UserID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "dashboard fetched", map[string]any{"summary": summary})
}

func (h *Handler) ListWorkerRequests(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	items, err := h.Module().WorkConnect.ListWorkerRequests(r.Context(), principal.UserID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "requests fetched", map[string]any{"requests": items})
}

func (h *Handler) GetWorkerRequest(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	requestID, err := parseIDParam(r, "requestID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid request id"))
		return
	}

	request, err := h.Module().WorkConnect.GetServiceRequestByID(r.Context(), requestID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	workerProfileID, isWorker, err := h.Module().WorkConnect.GetWorkerProfileInfo(r.Context(), principal.UserID)
	if err != nil || !isWorker {
		response.SendErrorResponse(w, r, apperrors.ErrForbidden)
		return
	}

	if workerProfileID != request.WorkerID {
		response.SendErrorResponse(w, r, apperrors.ErrForbidden)
		return
	}

	customer, err := h.Module().WorkConnect.GetUserByID(r.Context(), request.CustomerID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "request fetched", map[string]any{"request": request, "customer": customer})
}

func (h *Handler) WorkerDecision(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	requestID, err := parseIDParam(r, "requestID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid request id"))
		return
	}

	var req dto.WorkerDecisionRequest
	if err = decodeAndValidate(r, &req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	item, err := h.Module().WorkConnect.WorkerDecision(r.Context(), principal.UserID, requestID, req)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "request updated", map[string]any{"request": item})
}

func (h *Handler) StartWorkerRequest(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	requestID, err := parseIDParam(r, "requestID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid request id"))
		return
	}

	item, err := h.Module().WorkConnect.StartWorkerRequest(r.Context(), principal.UserID, requestID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "request updated", map[string]any{"request": item})
}

func (h *Handler) CompleteWorkerRequest(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	requestID, err := parseIDParam(r, "requestID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid request id"))
		return
	}

	item, err := h.Module().WorkConnect.CompleteWorkerRequest(r.Context(), principal.UserID, requestID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "request completed", map[string]any{"request": item})
}

func (h *Handler) ConfirmCustomerRequest(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	requestID, err := parseIDParam(r, "requestID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid request id"))
		return
	}

	item, err := h.Module().WorkConnect.ConfirmCustomerRequest(r.Context(), principal.UserID, requestID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "request updated", map[string]any{"request": item})
}

func (h *Handler) CancelCustomerRequest(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	requestID, err := parseIDParam(r, "requestID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid request id"))
		return
	}

	item, err := h.Module().WorkConnect.CancelCustomerRequest(r.Context(), principal.UserID, requestID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "request updated", map[string]any{"request": item})
}

func (h *Handler) WorkerAvailability(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	var req dto.UpdateAvailabilityRequest
	if err := decodeAndValidate(r, &req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	if err := h.Module().WorkConnect.UpdateWorkerAvailability(r.Context(), principal.UserID, req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "availability updated", nil)
}

func (h *Handler) WorkerDashboard(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	summary, err := h.Module().WorkConnect.WorkerDashboard(r.Context(), principal.UserID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "dashboard fetched", map[string]any{"summary": summary})
}

func (h *Handler) AdminDashboard(w nethttp.ResponseWriter, r *nethttp.Request) {
	summary, err := h.Module().WorkConnect.AdminDashboard(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "dashboard fetched", map[string]any{"summary": summary})
}

func (h *Handler) AdminUsers(w nethttp.ResponseWriter, r *nethttp.Request) {
	users, err := h.Module().WorkConnect.ListUsers(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}
	response.SendSuccessResponse(w, r, nethttp.StatusOK, "users fetched", map[string]any{"users": users})
}

func (h *Handler) AdminRequests(w nethttp.ResponseWriter, r *nethttp.Request) {
	requests, err := h.Module().WorkConnect.ListAllServiceRequests(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}
	response.SendSuccessResponse(w, r, nethttp.StatusOK, "requests fetched", map[string]any{"requests": requests})
}

func (h *Handler) PendingWorkers(w nethttp.ResponseWriter, r *nethttp.Request) {
	workers, err := h.Module().WorkConnect.PendingWorkerVerifications(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "workers fetched", map[string]any{"workers": workers})
}

func (h *Handler) ListWorkerDocuments(w nethttp.ResponseWriter, r *nethttp.Request) {
	workerID, err := parseIDParam(r, "workerID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid worker id"))
		return
	}

	documents, err := h.Module().WorkConnect.ListWorkerDocuments(r.Context(), workerID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "worker documents fetched", map[string]any{"documents": documents})
}

func (h *Handler) UploadWorkerDocument(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	var req dto.UploadWorkerDocumentRequest
	if err = decodeAndValidate(r, &req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	if err = h.Module().WorkConnect.UploadWorkerDocument(r.Context(), principal.UserID, req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "document uploaded", nil)
}

func (h *Handler) SubmitWorkerVerification(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	if err = h.Module().WorkConnect.SubmitWorkerVerification(r.Context(), principal.UserID); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "verification submitted", nil)
}

func (h *Handler) VerifyWorker(w nethttp.ResponseWriter, r *nethttp.Request) {
	workerID, err := parseIDParam(r, "workerID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid worker id"))
		return
	}

	if err = h.Module().WorkConnect.VerifyWorker(r.Context(), workerID, true); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "worker verified", nil)
}

func (h *Handler) ListMessageConversations(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	items, err := h.Module().WorkConnect.ListMessageConversations(r.Context(), principal.UserID)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "conversations fetched", map[string]any{"conversations": items})
}

func (h *Handler) ListMessagesByRequest(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	requestID, err := parseIDParam(r, "requestID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid request id"))
		return
	}

	limit := 50
	if limitRaw := r.URL.Query().Get("limit"); limitRaw != "" {
		parsedLimit, parseErr := strconv.Atoi(limitRaw)
		if parseErr != nil {
			response.SendErrorResponse(w, r, stderrs.New("invalid limit"))
			return
		}
		limit = parsedLimit
	}

	var beforeID int64
	if beforeRaw := r.URL.Query().Get("beforeId"); beforeRaw != "" {
		parsedBeforeID, parseErr := strconv.ParseInt(beforeRaw, 10, 64)
		if parseErr != nil {
			response.SendErrorResponse(w, r, stderrs.New("invalid beforeId"))
			return
		}
		beforeID = parsedBeforeID
	}

	items, err := h.Module().WorkConnect.ListMessagesByRequest(r.Context(), principal.UserID, requestID, dto.ListMessagesQuery{
		Limit:    limit,
		BeforeID: beforeID,
	})
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusOK, "messages fetched", map[string]any{"messages": items})
}

func (h *Handler) SendMessage(w nethttp.ResponseWriter, r *nethttp.Request) {
	principal, err := h.requirePrincipal(r.Context())
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	requestID, err := parseIDParam(r, "requestID")
	if err != nil {
		response.SendErrorResponse(w, r, stderrs.New("invalid request id"))
		return
	}

	var req dto.SendMessageRequest
	if err = decodeAndValidate(r, &req); err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	item, err := h.Module().WorkConnect.SendMessage(r.Context(), principal.UserID, requestID, req)
	if err != nil {
		response.SendErrorResponse(w, r, err)
		return
	}

	response.SendSuccessResponse(w, r, nethttp.StatusCreated, "message sent", map[string]any{"message": item})
}

func parseIDParam(r *nethttp.Request, paramName string) (int64, error) {
	return strconv.ParseInt(chi.URLParam(r, paramName), 10, 64)
}

func decodeAndValidate(r *nethttp.Request, dst any) error {
	if err := json.NewDecoder(r.Body).Decode(dst); err != nil {
		return stderrs.New("invalid payload")
	}

	req, ok := dst.(validatable)
	if !ok {
		return nil
	}

	return req.Validate()
}

func (h *Handler) requirePrincipal(ctx context.Context) (user.AuthPrincipal, error) {
	principal, ok := middleware.PrincipalFromContext(ctx)
	if !ok {
		return user.AuthPrincipal{}, apperrors.ErrUnauthorized
	}
	return principal, nil
}

func (h *Handler) authPayload(ctx context.Context, token string, user db.User) map[string]any {
	data := map[string]any{
		"token": token,
		"user":  user,
	}

	if user.Role == db.RoleWorker {
		if workerProfileID, _, err := h.Module().WorkConnect.GetWorkerProfileInfo(ctx, user.ID); err == nil {
			data["workerProfileId"] = workerProfileID
		}
	}

	return data
}

func (h *Handler) profilePayload(ctx context.Context, user db.User) map[string]any {
	data := map[string]any{"user": user}

	if user.Role == db.RoleWorker {
		if workerProfileID, _, err := h.Module().WorkConnect.GetWorkerProfileInfo(ctx, user.ID); err == nil {
			data["workerProfileId"] = workerProfileID
		}
	}

	return data
}

// func (h *Handler) writeError(w nethttp.ResponseWriter, r *nethttp.Request, err error) {
// 	response.SendErrorResponse(w, r, err)
// }