# openapi.yaml
openapi: 3.0.0
info:
  title: WorkConnect API
  description: Contract-first API specification for WorkConnect marketplace platform
  version: 1.0.0
  contact:
    name: WorkConnect Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: http://{host}:{port}/api/v1
    variables:
      host:
        default: localhost
      port:
        default: "8080"
paths:
...
  /workers:
    get:
      operationId: listWorkers
      tags:
        - Workers
      summary: List all workers with optional filters
      parameters:
        - name: category
          in: query
          description: Filter by service category
          schema:
            type: string
        - name: city
          in: query
          description: Filter by city
          schema:
            type: string
        - name: q
          in: query
          description: Search by name or skill
          schema:
            type: string
        - name: sort
          in: query
          description: Sort order
          schema:
            type: string
      responses:
        "200":
          description: Workers list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkersListResponse"
  /workers/{workerID}:
    get:
      operationId: getWorkerProfile
      tags:
        - Workers
      summary: Get detailed worker profile
      parameters:
        - name: workerID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Worker profile with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerDetailResponse"

  # ============================================================
  # AUTH ENDPOINTS
  # ============================================================
...
  /auth/me:
    get:
      operationId: getProfile
      tags:
        - Auth
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfileResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ============================================================
  # CUSTOMER ENDPOINTS
  # ============================================================
  /customer/requests:
    post:
      operationId: createCustomerRequest
      tags:
        - Customer
      summary: Create new service request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceRequest"
      responses:
        "201":
          description: Service request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: listCustomerRequests
      tags:
        - Customer
      summary: List customer's service requests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of customer's requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerRequestsListResponse"
  /customer/requests/{requestID}:
    get:
      operationId: getCustomerRequest
      tags:
        - Customer
      summary: Get specific customer request with worker details
      security:
        - bearerAuth: []
      parameters:
        - name: requestID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Request details with worker info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerRequestDetailResponse"
  /customer/requests/{requestID}/confirm:
    patch:
      operationId: confirmCustomerRequest
      tags:
        - Customer
      summary: Confirm worker for service request
      security:
        - bearerAuth: []
      parameters:
        - name: requestID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Request confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestResponse"
  /customer/requests/{requestID}/cancel:
    patch:
      operationId: cancelCustomerRequest
      tags:
        - Customer
      summary: Cancel service request
      security:
        - bearerAuth: []
      parameters:
        - name: requestID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Request cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestResponse"
  /customer/requests/{requestID}/review:
    post:
      operationId: submitCustomerReview
      tags:
        - Customer
      summary: Submit review for completed request
      security:
        - bearerAuth: []
      parameters:
        - name: requestID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitReviewRequest"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
  /customer/requests/{requestID}/payments/initiate:
    post:
      operationId: initiateCustomerPayment
      tags:
        - Customer
      summary: Initiate payment for service request
      security:
        - bearerAuth: []
      parameters:
        - name: requestID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiatePaymentRequest"
      responses:
        "201":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
  /customer/dashboard:
    get:
      operationId: customerDashboard
      tags:
        - Customer
      summary: Get customer dashboard summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDashboardResponse"

  # ============================================================
  # WORKER ENDPOINTS
  # ============================================================
  /worker/requests:
    get:
      operationId: listWorkerRequests
      tags:
        - Worker
      summary: List requests for worker
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of worker's requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerRequestsListResponse"
  /worker/requests/{requestID}:
    get:
      operationId: getWorkerRequest
      tags:
        - Worker
      summary: Get worker request with customer details
      security:
        - bearerAuth: []
      parameters:
        - name: requestID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Request details with customer info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerRequestDetailResponse"
  /worker/requests/{requestID}/start:
    patch:
      operationId: startWorkerRequest
      tags:
        - Worker
      summary: Mark request as started
      security:
        - bearerAuth: []
      parameters:
        - name: requestID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Request started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestResponse"
  /worker/requests/{requestID}/decision:
    patch:
      operationId: workerDecision
      tags:
        - Worker
      summary: Accept or reject service request
      security:
        - bearerAuth: []
      parameters:
        - name: requestID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkerDecisionRequest"
      responses:
        "200":
          description: Decision recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestResponse"
  /worker/requests/{requestID}/complete:
    patch:
      operationId: completeWorkerRequest
      tags:
        - Worker
      summary: Mark request as completed
      security:
        - bearerAuth: []
      parameters:
        - name: requestID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Request completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestResponse"
  /worker/availability:
    patch:
      operationId: updateWorkerAvailability
      tags:
        - Worker
      summary: Update worker availability status
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvailabilityRequest"
      responses:
        "200":
          description: Availability updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
  /worker/dashboard:
    get:
      operationId: workerDashboard
      tags:
        - Worker
      summary: Get worker dashboard summary
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerDashboardResponse"

  # ============================================================
  # MESSAGING ENDPOINTS
  # ============================================================
  /messages/conversations:
    get:
      operationId: listMessageConversations
      tags:
        - Messages
      summary: List message conversations for user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationsListResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register
  schemas:
...
    ProfileResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            workerProfileId:
              type: integer
              format: int64
              nullable: true
    WorkersListResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            workers:
              type: array
              items:
                $ref: "#/components/schemas/WorkerCard"
    WorkerDetailResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            worker:
              $ref: "#/components/schemas/WorkerDetails"
    ServiceRequestResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        data:
          type: object
          properties:
            request:
              $ref: "#/components/schemas/ServiceRequest"
...